pyproject.toml
[tool.poetry]
name = "speshelse"
version = "0.1.0"
description = ""
authors = ["Magne F. Myhren <mfm@ssb.no>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
ssb-klass-python = "^0.0.11"
cx-oracle = "^8.3.0"
openpyxl = "^3.1.2"
xlsxwriter = "^3.1.3"
python-dotenv = "^1.0.0"
tabulate = "^0.9.0"
numpy = "^2.2.0"
itables = "^2.2.4"
ssb-fagfunksjoner = "^1.0.7"
ssb-vaskify = "^0.0.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.15.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
